+++ androidApp/src/main/java/com/example/kotlin_social/android/post/PostDetailScreen.kt
@@ -45,7 +45,7 @@ fun PostDetailScreen(
     postUiState: PostUiState,
     commentsUiState: CommentsUiState,
     onCommentMoreIconClick: (Comment) -> Unit,
-    onProfileClick: (Long) -> Unit,
+    onProfileClick: (Int) -> Unit,
     onAddCommentClick: () -> Unit,
     fetchData: () -> Unit
 ){
